requirements.txt
absl-py==0.1.13
astor==0.6.2
azure-cognitiveservices-nspkg==2.0.0
azure-cognitiveservices-vision-customvision==0.1.0
azure-cognitiveservices-vision-nspkg==2.0.0
azure-common==1.1.9
azure-nspkg==2.0.0
bleach==1.5.0
certifi==2018.1.18
chardet==3.0.4
gast==0.2.0
grpcio==1.11.0
html5lib==0.9999999
idna==2.6
isodate==0.6.0
Markdown==2.6.11
msrest==0.4.27
numpy==1.14.2
oauthlib==2.0.7
pandas==0.22.0
Pillow==5.1.0
protobuf==3.5.2.post1
python-dateutil==2.7.2
pytz==2018.4
requests==2.18.4
requests-oauthlib==0.8.0
scipy==1.0.1
six==1.11.0
tensorboard==1.7.0
tensorflow==1.7.0
termcolor==1.1.0
urllib3==1.22
Werkzeug==0.14.1
